fix: Move close modal saga to the modal sagas
