+++ webapp/src/modules/modal/sagas.ts
@@ -2,11 +2,18 @@ import { LOCATION_CHANGE } from 'connected-react-router'
 import { delay, put, select, takeEvery } from 'redux-saga/effects'
 import { ModalState } from 'decentraland-dapps/dist/modules/modal/reducer'
 import { getOpenModals } from 'decentraland-dapps/dist/modules/modal/selectors'
-import { closeAllModals } from './actions'
+import { closeAllModals, closeModal } from './actions'
+import { CLAIM_LAND_SUCCESS } from '../rental/actions'
 
 export function* modalSaga() {
   yield takeEvery(LOCATION_CHANGE, handleLocationChange)
-  yield takeEvery([/** Actions that should trigger all modals to be closed */], handleCloseAllModals)
+  yield takeEvery(CLAIM_LAND_SUCCESS, handleClaimLandSuccess)
+  yield takeEvery(
+    [
+      /** Actions that should trigger all modals to be closed */
+    ],
+    handleCloseAllModals
+  )
 }
 
 function* handleLocationChange() {
@@ -15,6 +22,10 @@ function* handleLocationChange() {
     yield delay(100)
     yield handleCloseAllModals()
   }
+}
+
+function* handleClaimLandSuccess() {
+  yield put(closeModal('ClaimLandModal'))
 }
 
 function* handleCloseAllModals() {
